The [documentation](https://github.com/rs/cors/blob/master/cors.go#L36) implies that there's a performance penalty for using wildcards, but the performance can be improved.

Rather than concatenating two strings to find their combined length, it's quicker to just add their lengths.

Before:
```
$ go test -bench BenchmarkWildcard
goos: darwin
goarch: amd64
pkg: github.com/rs/cors
BenchmarkWildcard/match-4               50000000                25.8 ns/op             0 B/op          0 allocs/op
BenchmarkWildcard/too_short-4            100000000               20.7 ns/op             0 B/op          0 allocs/op
```
After:
```
$ go test -bench BenchmarkWildcard
goos: darwin
goarch: amd64
pkg: github.com/rs/cors
BenchmarkWildcard/match-4               200000000                7.45 ns/op            0 B/op          0 allocs/op
BenchmarkWildcard/too_short-4           2000000000               1.21 ns/op            0 B/op          0 allocs/op
```